Quick commit to add the other types of calculated nodes that I missed.

+++ real_time/surface_src/java_src/Alice/src/org/roboclub/robobuggy/calculatedNodes/NodeCalculator.java
@@ -5,5 +5,5 @@ import org.roboclub.robobuggy.messages.BaseMessage;
 public interface NodeCalculator {
 	
 	//Should be a function which returns time based measurements
-	public BaseMessage calculator(int millis);
+	public BaseMessage calculator(long millis);
 }
